requirements.txt
aiohttp==3.11.11
matplotlib==3.10.0
numpy==2.2.0
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
seaborn==0.13.2
streamlit==1.41.1